requirements.txt
streamlit
openai
PyMuPDF
fpdf
tqdm
scikit-learn
faiss-cpu
python-dotenv
Pillow
langchain
langchain-openai
langchain-community
